requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pinecone-client==2.2.4
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
mangum==0.17.0 